% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{BART}
\alias{BART}
\title{Run the BART algorithm for supervised learning}
\usage{
BART(
  X_train,
  y_train,
  W_train = NULL,
  X_test = NULL,
  W_test = NULL,
  feature_types = rep(0, ncol(X_train)),
  variable_weights = rep(1/ncol(X_train), ncol(X_train)),
  cutpoint_grid_size = 100,
  tau_init = NULL,
  alpha = 0.95,
  beta = 2,
  min_samples_leaf = 5,
  leaf_model = 0,
  nu = 3,
  lambda = NULL,
  a_leaf = 3,
  b_leaf = NULL,
  q = 0.9,
  sigma2_init = NULL,
  num_trees = 100,
  num_gfr = 5,
  num_burnin = 0,
  num_mcmc = 100,
  sample_sigma = T,
  sample_tau = T,
  random_seed = -1
)
}
\arguments{
\item{X_train}{}

\item{y_train}{}

\item{W_train}{}

\item{X_test}{}

\item{W_test}{}

\item{feature_types}{}

\item{variable_weights}{}

\item{tau_init}{}

\item{alpha}{}

\item{beta}{}

\item{min_samples_leaf}{}

\item{leaf_model}{}

\item{nu}{}

\item{lambda}{}

\item{a_leaf}{}

\item{b_leaf}{}

\item{sigma2_init}{}

\item{num_trees}{}

\item{num_gfr}{}

\item{num_burnin}{}

\item{num_mcmc}{}

\item{sample_sigma}{}

\item{sample_tau}{}

\item{random_seed}{}
}
\value{
An instance of the \code{BARTModel} class, sampled according to the provided parameters
}
\description{
Run the BART algorithm for supervised learning
}
