% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{convertToJson}
\alias{convertToJson}
\title{Convert a model object (BCF, BART, etc...) to JSON}
\usage{
convertToJson(object, ...)
}
\value{
Object of type \code{CppJson} which can be saved to disk with the \code{save_file(filename)}
method
}
\description{
Convert a model object (BCF, BART, etc...) to JSON
}
