% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{saveToJsonFile}
\alias{saveToJsonFile}
\title{Convert a model object (BCF, BART, etc...) to JSON and save it to a file
with a json suffix named \code{filename}}
\usage{
saveToJsonFile(object, filename, ...)
}
\description{
Convert a model object (BCF, BART, etc...) to JSON and save it to a file
with a json suffix named \code{filename}
}
